+++ tests/exclusions/expected/Module/java/Interface.java
@@ -1,0 +1,7 @@
+package com.example.module.model;
+
+// An interface
+public interface Interface extends Object {
+  String getKind();
+  void setKind(String value);
+}