+++ tests/exclusions/expected/Module/java/ObjectSubInterfaceOneOf3.java
@@ -1,0 +1,4 @@
+package com.example.module.model;
+
+public interface ObjectSubInterfaceOneOf3 extends Object {
+}