+++ tests/exclusions/expected/Module/java/ObjectSubObject.java
@@ -1,0 +1,15 @@
+package com.example.module.model;
+
+import lombok.*;
+
+@Builder
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ObjectSubObject {
+  private String key;
+  private String value;
+}
+
+