+++ tests/exclusions/expected/Module2/java/Object2AdditionalProperties.java
@@ -1,0 +1,15 @@
+package com.example.module2.model;
+
+import lombok.*;
+
+@Builder
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Object2AdditionalProperties {
+  private String key;
+  private String value;
+}
+
+